pyproject.toml
[project]
name = "finreport-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.18.6",
    "browser-use>=0.11.2",
    "bs4>=0.0.2",
    "crawl4ai>=0.7.8",
    "dill>=0.4.0",
    "docx2pdf>=0.1.8",
    "efinance>=0.5.5.2",
    "fake-useragent>=2.2.0",
    "fastapi>=0.128.0",
    "google-search-results>=2.4.2",
    "jieba>=0.42.0",
    "json-repair>=0.55.0",
    "matplotlib>=3.10.8",
    "numpy>=1.24.0",
    "openai>=2.14.0",
    "pandas>=2.0.0",
    "pdfplumber>=0.11.8",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "seaborn>=0.13.2",
    "streamlit>=1.52.2",
    "tavily-python>=0.7.19",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]
